requirements.txt
flask>=3.0.0
google-genai>=1.0.0
gunicorn>=22.0.0